sct sbbr tests: fix memory map allocation
The BBTestMemoryMapTest() function retrieves the UEFI memory map. To do so,
it goes through the following steps:

1. Call GetMemoryMap() a first time to obtain the size needed to hold the
   memory map.
2. Call SctAllocatePool() with the size obtained at step #1, to allocate an
   array.
3. Call GetMemoryMap() a second time with the array allocated at step #2,
   to actually get the memory map.

The problem with this approach as mentioned in the UEFI specification is
that the memory allocation at step #2 may very well modify the UEFI memory
map. This means the size obtained during step #1 may not be sufficient to
contain the UEFI memory map and the call at step #3 may fail. This is
indeed the actual behaviour with U-Boot, which results in a "MemoryMap Not
Found" error and a systematic failure of the SbbrBootServices Tests.

The recommendation in the UEFI specification is to allocate a bit more
memory at step #2 than just what is necessary and this is the strategy,
which is used by the UEFI SCT in multiple places.

We fix the allocation in BBTestMemoryMapTest() using the same method as
the SCT: allocating 1KB more. While at it, we apply the same fix
prophylactically to SbbrAllocAndGetMemoryMap().

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>