sct-test: build_bbr.sh script unsets ARCH
Remove definition of environment variable ARCH at entry of build_bbr.sh.
The script build edk2-test and edk2 components that expect that when
ARCH is defined for an arm architecture, it is either ARM or AARCH64.
Because build_bbr.sh can be called from tools that build common open
source packages that expect ARCH=arm|aarch64, this situation brings
confusion that is prevented by discarding the variable. The script
still expects SCT_TARGET_ARCH to describe the target architecture.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>