Added support for SCRT suite run
Added support to run SCRT in the ACS image automation run.
The results shall be collected and stored automatically in the acs_results

SCRT is a supplement to SCT and is invoked under the EFI shell environment and used to validate
UEFI Runtime Services implementations for compliance to the UEFI Specification.
These tests are invoked after calling ExitBootServices().